pantalla de carrito
ya esta listo la pantalla de carrito con responsive
