+++ Frontend/src/Components/CarritoCompras/responsiveCarrito.css
@@ -1,0 +1,233 @@
+/* --- Media Queries --- */
+
+
+/* Desktop (1441px - 1921px) */
+@media (min-width: 1441px) and (max-width: 1921px) {
+    .carrito-container {
+      padding: 0px;
+    }
+  
+    .producto-item {
+      flex-direction: row;
+      align-items: center;
+      margin-bottom: 30px;
+    }
+  
+    .producto-info p {
+      font-size: 1.2rem;
+    }
+  
+    .cantidad-controles button {
+      padding: 8px 12px;
+      
+    }
+  
+    .boton-continuar {
+      font-size: 1.3rem;
+    }
+  
+    header h1 {
+      font-size: 2.2rem;
+    }
+  
+    .carrito-icon span {
+      font-size: 1.8rem;
+    }
+  
+    .producto-imagen img {
+      max-width: 100px;
+    }
+  }
+  
+  /* Laptops (1024px - 1440px) */
+  @media (min-width: 1024px) and (max-width: 1440px) {
+    .carrito-container {
+      padding: 0px;
+    }
+
+    header h1 {
+        font-size: 1.4rem;
+        transform: translateX(-5px);
+        
+    }
+  
+    .producto-item {
+      flex-direction: row;
+      align-items: center;
+    }
+  
+    .cantidad-controles button {
+      padding: 5px 8px;
+    }
+  
+    .boton-continuar {
+      font-size: 1.2rem;
+    }
+  }
+  
+  /* Tablets (768px - 1023px) */
+  @media (min-width: 768px) and (max-width: 1023px) {
+    .carrito-container {
+      padding: 0px;
+    }
+  
+    header h1 {
+      font-size: 1.5rem;
+      transform: translateX(-5px);
+    }
+  
+    .producto-item {
+      flex-direction: column;
+      align-items: center;
+    }
+  
+    .producto-imagen img {
+      max-width: 100px;
+    }
+  
+    .cantidad-controles button {
+      padding: 4px 8px;
+    }
+  
+    .boton-continuar {
+      font-size: 1.1rem;
+    }
+  }
+  
+/*pantallas entre 425px y 767px */
+  @media (min-width: 425px) and (max-width: 767px) {
+    .cart-overlay {
+        width: 170vw;
+    }
+    
+    .carrito-container {
+      padding: 0px;
+    }
+  
+    header h1 {
+      font-size: 1.4rem;
+      transform: translateX(-5px);
+      
+    }
+  
+    .producto-item {
+      flex-direction: column;
+      align-items: center;
+      margin-bottom: 15px;
+    }
+  
+    .producto-imagen img {
+      max-width: 120px;
+    }
+  
+    .cantidad-controles {
+      margin-top: 10px;
+    }
+  
+    .cantidad-controles button {
+      padding: 4px 6px;
+    }
+  
+    .boton-continuar {
+      font-size: 1.1rem;
+    }
+  }
+  
+  /* pantallas entre 424px y 375px */
+@media (max-width: 424px) and (min-width: 375px) {
+    .mi-carrito {
+      padding: 0px;
+    }
+    .cart-overlay {
+        width: 180vw;
+    }
+  
+    .carrito-container {
+      padding: 0px;
+    }
+  
+    header h1 {
+      font-size: 1.2rem;
+      transform: translateX(-5px);
+    }
+  
+    .producto-item {
+      flex-direction: column;
+      align-items: center;
+      margin-bottom: 8px;
+    }
+  
+    .producto-imagen img {
+      max-width: 90px;
+    }
+  
+    .cantidad-controles {
+      margin-top: 10px;
+    }
+  
+    .cantidad-controles button {
+      padding: 4px 6px;
+    }
+  
+    .boton-continuar {
+      font-size: 1rem;
+      padding: 8px;
+    }
+  
+    .carrito-icon span {
+      font-size: 1.2rem;
+    }
+  }
+  
+ /*pantallas entre 374px y 320px */
+@media (max-width: 374px) and (min-width: 320px) {
+    
+    .cart-overlay {
+        width: 210vw;
+    }
+
+    .mi-carrito {
+      padding: 0px;
+    }
+  
+    .carrito-container {
+      padding: 0px;
+    }
+  
+    header h1 {
+      font-size: 1.2rem;
+      transform: translateX(0px);
+    }
+  
+    .producto-item {
+      flex-direction: column;
+      align-items: center;
+      margin-bottom: 8px;
+    }
+  
+    .producto-imagen img {
+      max-width: 90px;
+    }
+  
+    .producto-info {
+      font-size: 1rem;
+    }
+  
+    .cantidad-controles {
+      margin-top: 8px;
+    }
+  
+    .cantidad-controles button {
+      padding: 4px 6px;
+    }
+  
+    .boton-continuar {
+      font-size: 1rem;
+      padding: 10px;
+    }
+  
+    .carrito-icon span {
+      font-size: 1.2rem;
+    }
+  }
+
