+++ Frontend/src/Pages/Login/Login.css
@@ -468,184 +468,3 @@ margin-left: 5px;
   }
   
 }
-
-/*RESPONSIVE*/
-
-@media screen and (min-width: 375px) and (max-width: 424px) {
-  /* Contenedor principal */
-  .login-container {
-    padding: 5% 8%; /* Espacio adecuado para pantallas pequeñas */
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start; /* Alineación vertical */
-    height: auto; /* Ajustar la altura según el contenido */
-    overflow-y: auto; /* Permitir desplazamiento si es necesario */
-  }
-
-  /* Tarjeta de login */
-  .login-card {
-    width: 100%; /* Aseguramos que ocupe todo el ancho */
-    max-width: 90%; /* No exceder el 90% del ancho de la pantalla */
-    min-height: auto; /* Ajustar la altura según el contenido */
-    padding: 15px; /* Espacio dentro de la tarjeta */
-    margin-top: 10px; /* Separación superior */
-    border-radius: 15px; /* Bordes redondeados */
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Sombra suave */
-    border-radius: 60px;
-  }
-
-  /* Título */
-  .login-card h1 {
-    font-size: 7vw; /* Reducir tamaño de fuente con base al ancho de la pantalla */
-    margin-bottom: 12%; /* Espacio adicional debajo del título */
-    text-align: center; /* Centrar el título */
-    transform: translateY(30%); /* Ajuste en la posición */
-  }
-
-  /* Botón de login */
-  .login-button {
-    padding: 3% 30%; /* Reducir tamaño del botón */
-    font-size: 14px; /* Ajustar tamaño de fuente */
-    width: auto; /* Ajustar al contenido */
-    text-align: center;
-    margin-top: 20px; /* Espacio entre botones */
-    margin-left: 30px;
-  }
-
-  /* Enlace de "olvidé mi contraseña" */
-  .forgot-password {
-    font-size: 0.9rem; /* Ajuste del tamaño de la fuente */
-    margin-top: 12px; /* Espacio entre el campo y el enlace */
-    text-align: center; /* Centrado del enlace */
-    transform: translateY(-50%);
-  }
-
-  /* Enlace de "olvidé mi contraseña" */
-  .forgot-password a {
-    font-size: 0.9rem; /* Ajustar tamaño de fuente */
-    margin-left: 80px;
-  }
-
-  .social-buttons {
-    margin-left: -5%;
-  }
-
-  .login-facebook {
-    margin-right: 15px;
-    margin-left: 145px;
-  }
-
-  /* Mensaje de error */
-  .login-error-message {
-    margin-bottom: 10px;
-    font-size: 12px;
-    text-align: center; /* Centrar el mensaje de error */
-    margin-left: 0; /* Eliminar margen izquierdo */
-  }
-
-  /* Flecha de retroceso */
-  .back-login-circle {
-    width: 25px;
-    height: 25px;
-    margin-top: -10px;
-    margin-left: -10px;
-  }
-
-  /* Asegurarse que los campos estén bien distribuidos */
-  .form-group {
-    margin-bottom: 15px; /* Espacio adecuado entre campos */
-  }
-}
-
-
-@media screen and (max-width: 374px) {
-  
-  .login-container {
-    padding: 5% 0%; 
-    display: flex;
-    overflow-y: auto; 
-  }
-
-  /* Tarjeta de login */
-  .login-card {
-    width: 100%; /* Ajustar al 100% del ancho */
-    max-width: 90%; /* Hacer la tarjeta un poco más pequeña para evitar amontonamiento */
-    min-height: auto; /* Ajustar la altura a su contenido */
-    padding: 20px; /* Espacio adecuado dentro de la tarjeta */
-    margin-top: 10px; /* Espacio superior para separación */
-    border-radius: 15px; /* Bordes redondeados */
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Sombra sutil */
-    border-radius: 60px;
-  }
-
-  /* Título */
-  .login-card h1 {
-    font-size: 10vw; /* Tamaño adecuado para pantallas pequeñas */
-    margin-bottom: 10%; /* Mayor espacio debajo del título */
-    text-align: center;
-    transform: translateY(-10%); /* Ajuste leve en la posición */
-  }
-
-  /* Botón de login */
-  .login-button {
-    padding: 3% 30%; /* Espacio adecuado para el botón */
-    font-size: 14px; /* Tamaño de fuente legible */
-    width: auto; /* Ajustar al tamaño del texto */
-    text-align: center;
-    margin-top: 20px; /* Espacio entre los botones */
-  }
-
-  
-  .forgot-password {
-    font-size: 0.8rem; /* Ajustar tamaño de fuente para ser legible */
-    margin-top: 15px; /* Ajustar espacio para no quedar tan cerca del campo */
-    text-align: center; /* Centrar el enlace */
-    transform: translateY(-50%);
- 
-    margin-right: -90px;
-  }
-
- 
-  .forgot-password a {
-    font-size: 0.9rem; /* Tamaño adecuado */
-    color: #4CAF50;
-    transform: translateY(10%);
-  }
-
-  /* Botones sociales */
-  .social-buttons {
-    margin-left: -8%;
-  }
-
-  /* Mensaje de error */
-  .login-error-message {
-    margin-bottom: 5px;
-    font-size: 10px;
-    text-align: center;
-    margin-right: 20px; 
-  }
-
-  .back-login-circle {
-    width: 30px;
-    height: 30px;
-    margin-top: 20px; 
-  }
-
-  .form-group {
-    margin-bottom: 20px; 
-  }
-  .back-login-circle {
-    width: 25px;
-    height: 25px;
-  }
-  .login-back-button {
-    position: absolute;
-    top: -5px;
-    left: 25px;
-  }
-  .login-facebook {
-    margin-right: 30px;
-    margin-left: 150px;
-  }
-
-}
