+++ Frontend/src/Pages/pantallaPrincipal/pantallaPrincipal.jsx
@@ -20,6 +20,36 @@ const TuptiPage = ({ carouselImages, categoryImages }) => {
   const [error, setError] = useState(null);
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
   const [isCartOpen, setIsCartOpen] = useState(false);
+  
+
+  const [productos, setProductos] = useState([
+    { id: 1, nombre: "Manzana", precio: 10.5, cantidad: 1, imagen: "https://via.placeholder.com/150" },
+    { id: 2, nombre: "Chocolate", precio: 20.0, cantidad: 2, imagen: "https://via.placeholder.com/150" },
+    { id: 3, nombre: "Producto 3", precio: 5.5, cantidad: 5, imagen: "https://via.placeholder.com/150" },
+    { id: 4, nombre: "Producto 4", precio: 2.0, cantidad: 2, imagen: "https://via.placeholder.com/150" },
+    { id: 5, nombre: "Producto 5", precio: 6.0, cantidad: 2, imagen: "https://via.placeholder.com/150" },
+    { id: 6, nombre: "Producto 6", precio: 1.0, cantidad: 3, imagen: "https://via.placeholder.com/150" },
+  ]);
+
+
+
+  // Función para eliminar un producto del carrito
+  const eliminarProducto = (productoId) => {
+    const productosActualizados = productos.filter(producto => producto.id !== productoId);
+    setProductos(productosActualizados);
+  };
+  
+  const actualizarCantidad = (id, cantidad) => {
+    setProductos(
+      productos.map((producto) =>
+        producto.id === id
+          ? { ...producto, cantidad: Math.max(1, producto.cantidad + cantidad) }
+          : producto
+      )
+    );
+  };
+  
+  
 
   const toggleCart = () => {
     setIsCartOpen(!isCartOpen);
@@ -251,6 +281,7 @@ const TuptiPage = ({ carouselImages, categoryImages }) => {
           </Link>
           <button className="header-cart-button" onClick={toggleCart}>
         🛒 Carrito
+        <span className="icons-cart-counter">{productos.length}</span>
       </button>
 
         </div>
@@ -264,9 +295,13 @@ const TuptiPage = ({ carouselImages, categoryImages }) => {
         <div className="cart-overlay">
           <div className="cart-container">
             <button className="close-cart-button" onClick={toggleCart}>
-              ✖ Cerrar
+              ✖ 
             </button>
-            <CarritoCompras />
+            <CarritoCompras 
+            productos={productos}
+            eliminarProducto={eliminarProducto}
+            actualizarCantidad={actualizarCantidad}
+          />
           </div>
         </div>
       )}
